Compilation of numerous modifications and additions:
- Redesigned and implemented code to allow for the implication of a
"infinite" grid, memory of living cell positions for display and
processing, along with general navigation.
	- ArrayList keeps track of living cells present within the active game
state.
		- In response to user input, move living tiles in opposite direction.
		- Only draw tiles if their current position falls within the drawn
grid area. (This might be unnecessarily complex).
		- Navigation controls include the use of "WASD" and the arrow keys. In
addition to this, the space bar can be used to centre the camera.

- Added additional features to the UI.
	- Extra buttons, including centring the camera (May be obsolete).
		- A slider to allow for adjustment of "Speed", allowing to navigate
across the board as precise or quickly as is needed.
		- Labels to keep track of potentially important information, including
the current speed value and index position of the camera upon the board.
			- This not only informs the user of their active position, but also
helps to present the fact that there is a greater "grid" present outside
of the visible area.

- Restructured the "processing/Evolution" code.
	- Each generation, update the "Processing array".
		- This is cleared each iteration, and then re-populated with the
current "living cells", and their immediate neighbours.
		- These are the only cells the board will ever need to worry about.
The rest are generated when necessary. (Such as for drawing)
	- Then loop through all cells currently present within the array,
calculating the number of living cells neighbouring them.
	- Then run the appropriate predicates as to determine whether the cell
will become alive and be added to the next iterations "living cell"
array.

- Many more modifications, both small and large were made, however at
present this update note will be cut short. Ideally, these updates will
be made in smaller compilations in future.